Lift handleSuggestedJurisdictions out of component
