+++ app/src/rjsf/widgets/SingleSelectAutoComplete.tsx
@@ -6,57 +6,14 @@ import { useContext, useEffect, useState } from 'react';
 import { WidgetProps } from '@rjsf/utils';
 import { useSelector } from 'utils/use_selector';
 import { nanoid } from '@reduxjs/toolkit';
+import handleSuggestedJurisdictions from 'rjsf/business-rules/handleSuggestedJurisdictions';
+
 // Custom type to support this widget
 export type AutoCompleteSelectOption = {
   label: string;
   value: any;
   title: any;
   suggested?: boolean;
-};
-
-/**
- * @desc Helper used on jurisdictions keys. Iterates the SuggestedJurisdictions state and flags matching keys as suggested, sorting them to the top.
- * @param suggestedJurisdictionsInState Jurisdictions suggested given a particular area.
- * @param enumOptions Select Options list.
- */
-const handleSuggestedJurisdictions = (
-  suggestedJurisdictionsInState: Record<string, any>[],
-  enumOptions: AutoCompleteSelectOption[]
-) => {
-  const suggestedJurisdictions = suggestedJurisdictionsInState ? [...suggestedJurisdictionsInState] : [];
-  const additionalEnumOptions: AutoCompleteSelectOption[] = [];
-  suggestedJurisdictions.forEach((jurisdiction) => {
-    if (jurisdiction.geojson) {
-      additionalEnumOptions.push({
-        label: jurisdiction.geojson.properties.type ?? null,
-        value: jurisdiction.geojson.properties.code_name ?? null,
-        title: jurisdiction.geojson.properties.name ?? null,
-        suggested: true
-      } as AutoCompleteSelectOption);
-    } else if (jurisdiction.properties) {
-      additionalEnumOptions.push({
-        label: jurisdiction.properties.type ?? null,
-        value: jurisdiction.properties.code_name ?? null,
-        title: jurisdiction.properties.name ?? null,
-        suggested: true
-      } as AutoCompleteSelectOption);
-    }
-  });
-  enumOptions.forEach((option, index) => {
-    additionalEnumOptions.forEach((additionalOption) => {
-      if (option.value === additionalOption.value) {
-        enumOptions[index].suggested = true;
-      }
-    });
-  });
-  enumOptions.sort((left, right) => {
-    if (left.hasOwnProperty('suggested')) {
-      return -1;
-    } else if (right.hasOwnProperty('suggested')) {
-      return 1;
-    }
-    return 0;
-  });
 };
 
 /**
